Change 'rem' fns, fix tests, begin implementation
This commit removes the bitree_reml() and bitree_remr() functions,
replacing them with a single function, bitree_rem() which will remove
the node passed to it. Also fixes a couple of small bugs in the tests
which could result in segmentation faults. Finally bitree_create()
now passes all tests, and bitree_destroy() has been implemented. It
still currently fails the tests, however, due to a dependency on
bitree_insl(), which has not yet been implemented.