Implement bitree_merge()
The documentation in src/bitree.h was updated to reflect the exact
behaviour of bitree_merge. The function was also implemented, and
it passes all of the tests, however I do not believe it to be free
of bugs. Extensive testing will be sure to follow soon, with
corresponding patches.