Update bitree_merge() to increase efficiency in cases 2 & 3
In cases 2 & 3, we only need to recursively update the `size' and
`root' parameters of all of the nodes in tree2. This patch does that,
so that the best case efficiency of bitree_merge() (case 2 & 3) is
now n, where n is the size of tree2. Worst case efficiency is still
O(m + n), where m is the size of tree1.