STABLE RELEASE
This commit marks the stability of all features of the library. All
tests are passing, all necessary features are implemented. The bugs
file has been deleted, as it is no longer necessary, and test.c and
bitree.c have been updated with working versions of the source and
test for bitree_nlevelorder().